requirements.txt
annotated-types==0.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
idna==3.8
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.2
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.82
ultralytics-thop==2.0.5
urllib3==2.2.2
websocket-client==1.8.0
websockets==13.0
